Add docker development environment
Why:

In order to allow new developers to get started on working on new features and to ensure a consistent development environment experience, we are adding a docker development configuration. This will create a containerized environment that will include node 10, will install the depenedencies needed for local development, and will run the application.

This change addresses the need by:

* **Adding a docker-compose.builder.yml file.** - This file will be used to establish the install process that will pull the latest node files.
* **Adding a docker-compose.yml file.** - This file creates the server, runs the application, and forwards local port 3000 to port 3000 on the container. It should also be noted that the run command overrides the host ip address to 0.0.0.0 rather than localhost, which allows port forwarding to work as expected.
* **Adding a makefile.** - The makefile is used to install the application, and includes creating the nodemodules volume used by docker.
* **Updating README.md file.** - Adding notes for developers that would like to use the docker development environment for development.